+++ actix-redis/CHANGES.md
@@ -1,11 +1,21 @@
 # Changes
 
-## [Unreleased]
+## unreleased
 
-* Added `cookie_http_only` functionality to RedisSession builder, setting this
+* Minimize `futures` dependency
+
+## [0.9.0-alpha.2]
+
+* Add `cookie_http_only` functionality to RedisSession builder, setting this
   to false allows JavaScript to access cookies. Defaults to true.
 
-* Changed type of parameter of ttl method to u32.
+* Change type of parameter of ttl method to u32.
+
+* Update `actix` to 0.10.0-alpha.3
+
+* Update `tokio-util` to 0.3
+
+* Minimum supported Rust version(MSRV) is now 1.40.0.
 
 ## [0.9.0-alpha.1]