+++ actix-web-httpauth/CHANGES.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased (for alpha version)
+
+* Minimum supported Rust version(MSRV) is now 1.40.0.
+
 ## [unreleased]
   - Update the `base64` dependency to 0.12
+  - AuthenticationError's status code is preserved when converting to a ResponseError
+  - Minimize `futures` dependency
 
 ## [0.4.1] - 2020-02-19
   - Move repository to actix-extras